[FLINK-4002] [py] Reuse utility functions in tests
