+++ flink-libraries/flink-python/src/test/python/org/apache/flink/python/api/test_main2.py
@@ -18,76 +18,10 @@
 ################################################################################
 from flink.plan.Environment import get_environment
 from flink.functions.MapFunction import MapFunction
-from flink.functions.MapPartitionFunction import MapPartitionFunction
 from flink.functions.CrossFunction import CrossFunction
 from flink.functions.JoinFunction import JoinFunction
 from flink.functions.CoGroupFunction import CoGroupFunction
-
-
-#Utilities
-class Id(MapFunction):
-    def map(self, value):
-        """
-        Simple map function to forward test results.
-
-        :param value: Input value.
-        :return: Forwarded value.
-        """
-        return value
-
-
-class Verify(MapPartitionFunction):
-    def __init__(self, expected, name):
-        super(Verify, self).__init__()
-        self.expected = expected
-        self.name = name
-
-    def map_partition(self, iterator, collector):
-        """
-        Compares elements in the expected values list against actual values in resulting DataSet.
-        
-        :param iterator: Iterator for the corresponding DataSet partition.
-        :param collector: Collector for the result records.
-        """
-        index = 0
-        for value in iterator:
-            try:
-                if value != self.expected[index]:
-                    raise Exception(self.name + " Test failed. Expected: " + str(self.expected[index]) + " Actual: " + str(value))
-            except IndexError:
-                raise Exception(self.name + " Test failed. Discrepancy in the number of elements between expected and actual values.")
-            index += 1
-        #collector.collect(self.name + " successful!")
-        if(index != len(self.expected)):
-            raise Exception(self.name + " Test failed. Discrepancy in the number of elements between expected and actual values.")
-
-
-class Verify2(MapPartitionFunction):
-    def __init__(self, expected, name):
-        super(Verify2, self).__init__()
-        self.expected = expected
-        self.name = name
-
-    def map_partition(self, iterator, collector):
-        """
-        Compares elements in the expected values list against actual values in resulting DataSet.
-
-        This function does not compare element by element, since for example in a Union order is not guaranteed.
-
-        Elements are removed from the expected values list for the whole DataSet. 
-        
-        :param iterator: Iterator for the corresponding DataSet partition.
-        :param collector: Collector for the result records.
-        """
-        for value in iterator:
-            try:
-                self.expected.remove(value)
-            except Exception:
-                raise Exception(self.name + " failed! Actual value " + str(value) + " not contained in expected values: " + str(self.expected))
-        #collector.collect(self.name + " successful!")
-        if self.expected:
-            raise Exception(self.name + " Test failed. Discrepancy in the number of elements between expected and actual values.")
-
+from utils import Verify, Verify2
 
 if __name__ == "__main__":
     env = get_environment()
